+++ app/scripts/models/file-info-model.js
@@ -12,7 +12,9 @@ var FileInfoModel = Backbone.Model.extend({
         editState: null,
         rev: null,
         syncDate: null,
-        openDate: null
+        openDate: null,
+        keyFileName: null,
+        keyFileHash: null
     },
 
     initialize: function(data, options) {