+++ app/scripts/const/timeouts.js
@@ -4,7 +4,8 @@ var Timeouts = {
     AutoSync: 30 * 1000 * 60,
     CopyTip: 1500,
     AutoHideHint: 3000,
-    FileChangeSync: 3000
+    FileChangeSync: 3000,
+    BeforeAutoLock: 300
 };
 
 module.exports = Timeouts;